==> data_engineer/01_data2db.ipynb <==
# imports
import datetime as dt
import pandas as pd

from de.controller.db.controller import get_db_engine

from de.loader.api_loader.dart import get_dart_api_loader, get_dart_fss_loader
from de.loader.api_loader.fdr import get_fdr_loader
from de.loader.api_loader.pykrx import get_pykrx_loader

from de.loader.static_loader.krx import KRX_LOADER

# CFG

## dart_api_cfg
DART_API_CFG = {
    "year": 2023,
    # (11011 : 연간보고서) / (11012 : 반기보고서) / (11013 : 1분기보고서) / (11014 : 3분기보고서)
    "reprt_code": 11014,
}

## fdr_cfg
FDR_CFG = {
    "start": "20200101",
    # "end": dt.date.today().strftime("%Y%m%d"),
    "end": (dt.date.today() - dt.timedelta(days=1)).strftime("%Y%m%d"),
}
## pykrx_cfg
PYKRX_CFG = {
    "start": "20200101",
    # "end": dt.date.today().strftime("%Y%m%d"),
    "end": (dt.date.today() - dt.timedelta(days=1)).strftime("%Y%m%d"),
}

db_engine = get_db_engine()

# Dart

## dart_fss
dart_fss_loader = get_dart_fss_loader()

## dart_fss : dart_info
dart_info_df = dart_fss_loader.get_info_df()
dart_info_df = dart_info_df[~(dart_info_df["stock_code"].isna())]
dart_info_df.to_sql(name="dart_info_df", con=db_engine, if_exists="replace", index=False)


## dart_api
dart_api_loader = get_dart_api_loader()

corp_codes = sorted(set(dart_info_df["corp_code"]))

### dart_api : dart_fundamental
dart_fundamental_df = dart_api_loader.get_corps_fundamental_df(
    corp_codes=corp_codes, year=DART_API_CFG["year"], reprt_code=DART_API_CFG["reprt_code"]
)

dart_fundamental_df.to_sql(name="dart_fundamental_df", con=db_engine, if_exists="replace", index=False)

# Fdr

## fdr
fdr_loader = get_fdr_loader()

stock_codes = sorted(set(dart_info_df["stock_code"]))

### fdr_info_df
fdr_info_df = fdr_loader.get_krx_df()
fdr_info_df.to_sql(name="fdr_info_df", con=db_engine, if_exists="replace", index=False)


### fdr_ohlcv_df
fdr_ohlcv_df = fdr_loader.get_stocks_ohlcv_df(StockCodes=stock_codes, start=FDR_CFG["start"], end=FDR_CFG["end"])
fdr_ohlcv_df.to_sql(name="fdr_ohlcv_df", con=db_engine, if_exists="replace", index=False)

# pykrx
pykrx_loader = get_pykrx_loader()

## pykrx_info_df
KOSPI_trader_df = pykrx_loader.get_stock_trader_df(StockCode="KOSPI", start=PYKRX_CFG["start"], end=PYKRX_CFG["end"])
KOSPI_trader_df["Market"] = "KOSPI"

KOSDAQ_trader_df = pykrx_loader.get_stock_trader_df(StockCode="KOSDAQ", start=PYKRX_CFG["start"], end=PYKRX_CFG["end"])
KOSDAQ_trader_df["Market"] = "KOSDAQ"

pykrx_info_df = pd.concat([KOSDAQ_trader_df, KOSPI_trader_df], axis=0)
pykrx_info_df.reset_index(inplace=True)
pykrx_info_df.to_sql(name="pykrx_info_df", con=db_engine, if_exists="replace", index=False)

# KRX
krx_loader = KRX_LOADER("./de/data")

## krx_info
krx_info_df = krx_loader.get_krx_info_df()
krx_info_df.to_sql(name="krx_info_df", con=db_engine, if_exists="replace", index=False)
