==> data_engineer/02_db2local.ipynb <==
import pandas as pd

from de.controller.db.controller import get_db_engine, DB_CONTROLLER
from de.processor.db.modifier import DB_MODIFIER

db_controller = DB_CONTROLLER(get_db_engine())
db_modifier = DB_MODIFIER()

# dart

## dart_fundamental
dart_fundamental_df = db_controller.get_dart_fundamental_df()
md_dart_fundamental_df = db_modifier.get_md_dart_fundamental_df(dart_fundamental_df)
md_dart_fundamental_df.to_csv("./de/data/dart_fundamental_df.csv")


## dart_info
dart_info_df = db_controller.get_dart_info_df()
md_dart_info_df = db_modifier.get_md_dart_info_df(dart_info_df)
md_dart_info_df.to_csv("./de/data/dart_info_df.csv")

# fdr

## fdr_info
fdr_info_df = db_controller.get_fdr_info_df()
md_fdr_info_df = db_modifier.get_md_fdr_info_df(fdr_info_df)
md_fdr_info_df.to_csv("./de/data/fdr_info_df.csv")

## fdr_ohlcv
fdr_ohlcv_df = db_controller.get_fdr_ohlcv_df()
md_fdr_ohlcv_df = db_modifier.get_md_fdr_ohlcv_df(fdr_ohlcv_df)
md_fdr_ohlcv_df.to_csv("./de/data/fdr_ohlcv_df.csv")

# pykrx

## pykrx_info
pykrx_info_df = db_controller.get_pykrx_info_df()
md_pykrx_info_df = db_modifier.get_md_pykrx_info_df(pykrx_info_df)
md_pykrx_info_df.to_csv("./de/data/pykrx_info_df.csv")

# krx

## krx_info
krx_info_df = db_controller.get_krx_info_df()
md_krx_info_df = db_modifier.get_md_krx_info_df(krx_info_df)
md_krx_info_df.to_csv("./de/data/krx_info_df.csv")
