==> data_engineer/06_order.ipynb <==
import datetime as dt
import pandas as pd

from de.controller.koreainvestment.controller import get_ki_broker, ORDER_CONTROLLER

CFG = {
    "BUYING": {
        "stocks": {"010950", "003490", "006660"},
    },
    "SELLING": {
        "upper_pct": 0.06,
        "lower_pct": 0.06 / 3,
    },
}


ki_broker = get_ki_broker()
order_controller = ORDER_CONTROLLER(ki_broker)

order_preprocessor = order_controller.order_preprocessor()
order_requester = order_controller.order_requester()

buying_df = order_preprocessor.get_buying_df(CFG["BUYING"]["stocks"])
buying_df

# """
# 시장가 매수
# """
# order_resp_dict = dict()

# for idx, row in buying_df.iterrows():
#     stockcode = row["StockCode"]
#     quantity = row["Quantity"]
#     order_resp = order_requester.request_market_buy_order(stockcode=stockcode, quantity=quantity)
#     order_resp_dict[stockcode] = order_resp

# """
# 시장가 매수 예약
# """

# buying_resv_resp_dict = dict()

# for idx, row in buying_df.iterrows():
#     StockCode = row["StockCode"]
#     Quantity = row["Quantity"]

#     buying_resv_resp = order_requester.request_market_buy_order_resv(
#         stockcode=StockCode,
#         quantity=Quantity,
#         resv_date=(dt.date.today() + dt.timedelta(days=1)),
#     )
#     buying_resv_resp_dict[StockCode] = buying_resv_resp

# """
# 예약 매도 주문 파악
# """
# open_sell_order_resv = ki_broker.fetch_open_order_resv(side="sell")
# rename_dict = {
#     "pdno": "StockCode",
#     "kor_item_shtn_name": "StockName",
#     "ord_rsvn_qty": "Quantity",
#     "ord_rsvn_unpr": "Price",
# }
# try:
#     open_sell_order_resv_df = pd.DataFrame(open_sell_order_resv["output"])
#     open_sell_order_resv_df.rename(columns=rename_dict, inplace=True)
#     open_sell_order_resv_df.loc[:, list(rename_dict.values())]
# except:
#     open_sell_order_resv_df = pd.DataFrame(columns=list(rename_dict.values()))
# open_sell_order_resv_df

# """
# 예약 매수 주문 파악
# """
# open_buy_order_resv = ki_broker.fetch_open_order_resv(side="buy")

# rename_dict = {
#     "pdno": "StockCode",
#     "kor_item_shtn_name": "StockName",
#     "ord_rsvn_qty": "Quantity",
#     "ord_rsvn_unpr": "Price",
# }
# try:
#     open_buy_order_resv_df = pd.DataFrame(open_buy_order_resv["output"])
#     open_buy_order_resv_df.rename(columns=rename_dict, inplace=True)
#     open_buy_order_resv_df.loc[:, list(rename_dict.values())]
# except:
#     open_buy_order_resv_df = pd.DataFrame(columns=list(rename_dict.values()))
# open_buy_order_resv_df

position_df = order_preprocessor.get_position_df()
position_df

# 예약 매도 주문 데이터 생성
resv_order_df = order_preprocessor.get_resv_order_df(position_df)
resv_order_df

# """
# 예약 주문 매도
#     - 익절
#     - 손절
# """

# order_resv_resp_dict = dict()

# for idx, row in resv_order_df.iterrows():
#     StockCode = row["StockCode"]
#     Quantity = row["Quantity"]
#     ResvUpperPrice = row["ResvUpperPrice"]
#     ResvLowerPrice = row["ResvLowerPrice"]

#     order_resv_lower_resp = order_requester.request_limit_sell_order_resv(
#         stockcode=StockCode,
#         price=ResvLowerPrice,
#         quantity=Quantity,
#         resv_date=(dt.date.today() + dt.timedelta(days=11)),
#     )

#     order_resv_upper_resp = order_requester.request_limit_sell_order_resv(
#         stockcode=StockCode,
#         price=ResvUpperPrice,
#         quantity=Quantity,
#         resv_date=(dt.date.today() + dt.timedelta(days=11)),
#     )

#     order_resv_resp_dict[StockCode] = (order_resv_lower_resp, order_resv_upper_resp)
