==> data_engineer/99_analysis_stock.ipynb <==
import datetime as dt
import pandas as pd

from de.processor.source.dart import DART_FUNDAMENTAL_PROCESSOR, DART_INFO_PROCESSOR
from de.processor.source.krx import KRX_INFO_PROCESSOR
from de.processor.source.fdr import FDR_INFO_PROCESSOR, FDR_OHLCV_PREPROCESSOR

# dart_fundamental
dart_fundamental_df = pd.read_csv("./de/data/dart_fundamental_df.csv", index_col=0)
dart_fundamental_processor = DART_FUNDAMENTAL_PROCESSOR(dart_fundamental_df)

# dart_info
dart_info_df = pd.read_csv("./de/data/dart_info_df.csv", index_col=0)
dart_info_processor = DART_INFO_PROCESSOR(dart_info_df)

# krx_info
krx_info_df = pd.read_csv("./de/data/krx_info_df.csv", index_col=0)
krx_info_processor = KRX_INFO_PROCESSOR(krx_info_df)

# fdr_info
fdr_info_df = pd.read_csv("./de/data/fdr_info_df.csv", index_col=0)
fdr_info_processor = FDR_INFO_PROCESSOR(fdr_info_df)

# fdr_ohlcv
fdr_ohlcv_df = pd.read_csv("./de/data/fdr_ohlcv_df.csv", index_col=0)
fdr_ohlcv_preprocessor = FDR_OHLCV_PREPROCESSOR(fdr_ohlcv_df)
ohlcv_df = fdr_ohlcv_preprocessor.get_pps_fdr_ohlcv_df()

class STOCK_INFO_PROCESOR:
    def __init__(
        self, fdr_info_processor, krx_info_processor, dart_info_processor, dart_fundamental_processor, ohlcv_df
    ):
        self.fdr_info_processor = fdr_info_processor
        self.krx_info_processor = krx_info_processor
        self.dart_info_processor = dart_info_processor
        self.dart_fundamental_processor = dart_fundamental_processor
        self.ohlcv_df = ohlcv_df

    @staticmethod
    def append_StockName(df, fdr_info_processor):
        df["StockName"] = df["StockCode"].map(fdr_info_processor.get_StockName_dict())
        return df

    @staticmethod
    def append_MarketSector(df, krx_info_processor):
        df["MarketSector"] = df["StockCode"].map(krx_info_processor.get_MarketSector_dict())
        return df

    @staticmethod
    def append_Product(df, dart_info_processor):
        df["Product"] = df["StockCode"].map(dart_info_processor.get_Product_dict())
        return df

    @staticmethod
    def append_bps(df, dart_fundamental_processor, fdr_info_processor):
        df["BPS"] = (
            df["StockCode"].map(dart_fundamental_processor.get_TotalAssets_dict())
            / df["StockCode"].map(fdr_info_processor.get_Shares_dict())
        ).astype(int)
        return df

    @staticmethod
    def append_eps(df, dart_fundamental_processor, fdr_info_processor):
        df["EPS"] = (
            df["StockCode"].map(dart_fundamental_processor.get_NetProfit_dict())
            / df["StockCode"].map(fdr_info_processor.get_Shares_dict())
        ).astype(int)
        return df

    @staticmethod
    def append_price(df, ohlcv_df):
        price_dict = ohlcv_df[ohlcv_df["Date"] == ohlcv_df["Date"].max()].set_index("StockCode")["Close"].to_dict()
        df["Price"] = df["StockCode"].map(price_dict)
        return df

    def get_stock_info_df(self, stockcodes):
        stock_info_df = pd.DataFrame(stockcodes, columns=["StockCode"])
        stock_info_df = self.append_StockName(stock_info_df, self.fdr_info_processor)
        stock_info_df = self.append_MarketSector(stock_info_df, self.krx_info_processor)
        stock_info_df = self.append_Product(stock_info_df, self.dart_info_processor)
        stock_info_df = self.append_eps(stock_info_df, self.dart_fundamental_processor, self.fdr_info_processor)
        stock_info_df = self.append_bps(stock_info_df, self.dart_fundamental_processor, self.fdr_info_processor)
        stock_info_df = self.append_price(stock_info_df, self.ohlcv_df)
        return stock_info_df

stock_info_procesor = STOCK_INFO_PROCESOR(
    fdr_info_processor, krx_info_processor, dart_info_processor, dart_fundamental_processor, ohlcv_df
)

f_ma_stockcodes = {'047050', '063170', '066970', '079810'}
f_ma_stock_info_df = stock_info_procesor.get_stock_info_df(stockcodes=f_ma_stockcodes)
f_ma_stock_info_df

f_lc_stockcodes = {'023960', '298040'}
f_lc_stock_info_df = stock_info_procesor.get_stock_info_df(stockcodes=f_lc_stockcodes)
f_lc_stock_info_df
